#410: Add GuideSceneContentsBuilder
